refactor: split utils package out to config subpackages
