+++ config/config_test.go
@@ -1,7 +1,9 @@
 package config
 
 import (
+	"io/ioutil"
 	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/joyent/containerpilot/tests/assert"
@@ -145,4 +147,77 @@ func TestCustomConfigControl(t *testing.T) {
 	assert.Equal(t, cfg.Control.SocketPath,
 		"/var/run/cp3-test.sock",
 		"expected '%v' for control.socket, but got '%v'")
+}
+
+func TestInvalidRenderConfigFileMissing(t *testing.T) {
+	err := RenderConfig("/xxxx", "-")
+	assert.Error(t, err,
+		"could not read config file: open /xxxx: no such file or directory")
+}
+
+func TestInvalidRenderConfigOutputMissing(t *testing.T) {
+	err := RenderConfig("./testdata/test.json5", "./xxxx/xxxx")
+	assert.Error(t, err,
+		"could not write config file: open ./xxxx/xxxx: no such file or directory")
+}
+
+func TestRenderConfigFileStdout(t *testing.T) {
+
+	// Render to file
+	defer os.Remove("testJSON.json")
+	if err := RenderConfig("./testdata/test.json5", "testJSON.json"); err != nil {
+		t.Fatalf("expected no error from renderConfigTemplate but got: %v", err)
+	}
+	if exists, err := fileExists("testJSON.json"); !exists || err != nil {
+		t.Errorf("expected file testJSON.json to exist.")
+	}
+
+	// Render to stdout
+	fname := filepath.Join(os.TempDir(), "stdout")
+	temp, _ := os.Create(fname)
+	old := os.Stdout
+	os.Stdout = temp
+	if err := RenderConfig("./testdata/test.json5", "-"); err != nil {
+		t.Fatalf("expected no error from renderConfigTemplate but got: %v", err)
+	}
+	temp.Close()
+	os.Stdout = old
+
+	renderedOut, _ := ioutil.ReadFile(fname)
+	renderedFile, _ := ioutil.ReadFile("testJSON.json")
+	if string(renderedOut) != string(renderedFile) {
+		t.Fatalf("expected the rendered file and stdout to be identical")
+	}
+}
+
+func TestRenderedConfigIsParseable(t *testing.T) {
+
+	var testJSON = `{
+	"consul": "consul:8500",
+	watches: [{"name": "upstreamA{{.TESTRENDERCONFIGISPARSEABLE}}", "interval": 11}]}`
+
+	os.Setenv("TESTRENDERCONFIGISPARSEABLE", "-ok")
+	template, _ := renderConfigTemplate([]byte(testJSON))
+	config, err := newConfig(template)
+	if err != nil {
+		t.Fatalf("unexpected error in LoadConfig: %v", err)
+	}
+	name := config.Watches[0].Name
+	if name != "watch.upstreamA-ok" {
+		t.Fatalf("expected Watches[0] name to be upstreamA-ok but got %s", name)
+	}
+}
+
+// ----------------------------------------------------
+// test helpers
+
+func fileExists(path string) (bool, error) {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true, nil
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return true, err
 }
