+++ config/decoding/decode_test.go
@@ -1,4 +1,4 @@
-package utils
+package decoding
 
 import (
 	"reflect"
