+++ config/services/ips_test.go
@@ -1,4 +1,4 @@
-package utils
+package services
 
 import (
 	"fmt"
