+++ config/services/names.go
@@ -1,4 +1,4 @@
-package utils
+package services
 
 import (
 	"fmt"
