+++ config/services/names_test.go
@@ -1,4 +1,4 @@
-package utils
+package services
 
 import (
 	"testing"
