+++ config/templating/template.go
@@ -1,4 +1,4 @@
-package config
+package templating
 
 import (
 	"bytes"
