+++ config/timing/duration.go
@@ -1,4 +1,4 @@
-package utils
+package timing
 
 import (
 	"fmt"
