+++ config/timing/duration_test.go
@@ -1,4 +1,4 @@
-package utils
+package timing
 
 import (
 	"errors"
